Reorganize core header files
