+++ include/clement/clement.hpp
@@ -1,16 +1,16 @@
 #ifndef CLEMENT_CLEMENT_HPP
 #define CLEMENT_CLEMENT_HPP
 
-#include "constants.hpp"
-#include "http_server.hpp"
-#include "logger.hpp"
-#include "path.hpp"
-#include "request.hpp"
-#include "response_builder.hpp"
-#include "route.hpp"
-#include "router.hpp"
-#include "utility.hpp"
-#include "writer.hpp"
+#include "clement/core/constants.hpp"
+#include "clement/core/http_server.hpp"
+#include "clement/core/logger.hpp"
+#include "clement/core/path.hpp"
+#include "clement/core/request.hpp"
+#include "clement/core/response_builder.hpp"
+#include "clement/core/route.hpp"
+#include "clement/core/router.hpp"
+#include "clement/core/utility.hpp"
+#include "clement/core/writer.hpp"
 #include <boost/beast/http.hpp>
 
 namespace clement {
