+++ include/clement/core/utility.hpp
@@ -1,0 +1,9 @@
+#ifndef CLEMENT_UTILITY_HPP
+#define CLEMENT_UTILITY_HPP
+
+#include "clement/utility/base64.hpp"
+#include "clement/utility/form_urlencoded.hpp"
+#include "clement/utility/generate_string.hpp"
+#include "clement/utility/url_utility.hpp"
+
+#endif
