--- include/clement/utility.hpp
@@ -1,9 +1,0 @@
-#ifndef CLEMENT_UTILITY_HPP
-#define CLEMENT_UTILITY_HPP
-
-#include "utility/base64.hpp"
-#include "utility/form_urlencoded.hpp"
-#include "utility/generate_string.hpp"
-#include "utility/url_utility.hpp"
-
-#endif
