--- test/src/unit_tests.cpp
@@ -1,5 +1,0 @@
-#define CATCH_CONFIG_MAIN
-
-#include "utility_test.hpp"
-#include "path_test.hpp"
-#include "request_test.hpp"
